+++ sdk/resourcemanager/network/armnetwork/CHANGELOG.md
@@ -1,10 +1,31 @@
 # Release History
 
-## 3.1.0-beta.1 (2023-06-12)
+## 4.0.0 (2023-07-14)
+### Breaking Changes
+
+- `ApplicationGatewayCustomErrorStatusCodeHTTPStatus499` from enum `ApplicationGatewayCustomErrorStatusCode` has been removed
 
 ### Features Added
 
-- Support for test fakes and OpenTelemetry trace spans.
+- New value `ApplicationGatewaySKUNameBasic` added to enum type `ApplicationGatewaySKUName`
+- New value `ApplicationGatewayTierBasic` added to enum type `ApplicationGatewayTier`
+- New enum type `AdminState` with values `AdminStateDisabled`, `AdminStateEnabled`
+- New enum type `SyncMode` with values `SyncModeAutomatic`, `SyncModeManual`
+- New function `*LoadBalancersClient.MigrateToIPBased(context.Context, string, string, *LoadBalancersClientMigrateToIPBasedOptions) (LoadBalancersClientMigrateToIPBasedResponse, error)`
+- New struct `MigrateLoadBalancerToIPBasedRequest`
+- New struct `MigratedPools`
+- New field `ResourceGUID` in struct `AdminPropertiesFormat`
+- New field `ResourceGUID` in struct `AdminRuleCollectionPropertiesFormat`
+- New field `HTTPListeners` in struct `ApplicationGatewayFrontendIPConfigurationPropertiesFormat`
+- New field `DefaultPredefinedSSLPolicy` in struct `ApplicationGatewayPropertiesFormat`
+- New field `SyncMode` in struct `BackendAddressPoolPropertiesFormat`
+- New field `ResourceGUID` in struct `ConnectivityConfigurationProperties`
+- New field `ResourceGUID` in struct `DefaultAdminPropertiesFormat`
+- New field `ResourceGUID` in struct `GroupProperties`
+- New field `ResourceGUID` in struct `ManagerProperties`
+- New field `ResourceGUID` in struct `SecurityAdminConfigurationPropertiesFormat`
+- New field `AdminState` in struct `VirtualNetworkGatewayPropertiesFormat`
+
 
 ## 3.0.0 (2023-05-26)
 ### Breaking Changes